+++ materias/templatetags/materias_tags.py
@@ -14,3 +14,11 @@ def restantes_hasta(it, cantidad):
 def cargo_nombre_largo(cargo_dedicacion):
     cd = CargoDedicacion[cargo_dedicacion]
     return cd.value
+
+@register.filter
+def bien_o_mal(a, b):
+    if a == b:
+        return "bien"
+    if a > b:
+        return "maso"
+    return "mal"
